Run e2e tests locally

+++ playwright.global-teardown.ts
@@ -1,4 +1,6 @@
 import { execSync } from 'child_process';
 export default async function globalTeardown() {
-  execSync('docker compose down', { stdio: 'inherit' });
+  if (process.env.NO_DOCKER !== '1') {
+    execSync('docker compose down', { stdio: 'inherit' });
+  }
 }
